// lib/types.ts
// {
//     "id": "416624c4-45b0-40ec-ab25-fc99a0f85f6f",
//     "created_at": "2023-07-17T09:50:54.060014+00:00",
//     "name": "Digi Convertor",
//     "desc": "Small school project that allows you to convert between multiple digit formats",
//     "tags": "nms small animation transition typing dark mode",
//     "github_url": "https://github.com/Clupai8o0/digi-convertor-app",
//     "website_url": "https://digi-convertor.vercel.app/",
//     "image_url": "0616b8fa-074e-4ef9-9c0d-887789a34ae0.jpg"
// }

export interface Link {
	id: string;
	created_at: string;
	name: string;
	desc: string;
	tags: string;
	github_url: string;
	website_url: string;
	image_url: string;
}